feat: MQTT and pull vehicle info parallel processing
First MQTT implementation was inserting every vehicle row to DB
separately and pull reception (currently e.g. Tampere data) was
switched off. Now pull reception switched back on and MQTT processing
changed so that new vehicle rows are first inserted into an array (as
in pull reception) and the push_array is inserted to DB while
processing pull data (currently every 30 secs). This had two
implications: 1) Used a global array and a lock so that the two
threads (MQTT reception and pull processing) don't collide and 2)
It seems that sometimes in MQTT data the same vehicle can occur more
than once with the same timestamp. It is a database unique constraint
and there is a filtering rule against it, but the DB still borks if
a duplicate comes in the same insert. So, added an extra loop to filter
duplicate entries from a single insert.