doc: Explained how to retrieve cancelled trips, if decide to implement
Currently using ServiceAlerts, which will be deprecated in the future
in favor of DisruptionInfo. It appears that neither ServiceAlerts nor
DisruptionInfo is informing about cancelled departures (nothing
captured for a long time, but may be a retrieval issue - check).
Cancelled departures can be queried separately, but will require
some attention if used in parallel with ServiceAlerts/DisruptionInfo,
because duplicate checking is currently based in alertId, which is
not available from cancelledTripTimes.
Anyway, added a GraphQLQuery for cancelledTripTimes in comments as
an example of how it could look like.